requirements.txt
ads1015==1.0.0
colorama==0.4.6
gpiod==2.2.1
gpiodevice==0.0.5
gpiozero==2.0.1
icm20948==1.0.0
matrix11x7==0.0.1
more-itertools==8.10.0
pigpio==1.78
pimoroni-ioexpander==1.0.1
pmw3901==1.0.0
psutil==5.9.4
pygame==2.6.1
pyserial==3.5
pytest==7.2.1
python-daemon==3.1.2
PyYAML==6.0.2
rgbmatrix5x5==0.0.3
rpi-ws281x==5.0.0
rshell==0.0.34
smbus2==0.4.2
spidev==3.5
toml==0.10.2
vl53l5cx-ctypes==0.0.3
